7x on the whole suite, and Eq. 80 solved exactly instead of bisected
From the performance audit. Two hard constraints, both met: no new dependency
(`math` and `bisect` are stdlib, and `tests/test_imports.py` enforces the rest)
and no accuracy loss.

| workload | before | after |
|---|---|---|
| held 400 lbm/hr trim | 205.8 us/frame | 31.5 |
| Figure 9 accel | 236.4 | 37.4 |
| Figure 10 chop | 522.2 | 48.8 |
| closed loop | 515.0 | 95.5 |
| `trim.solve` | 5.77 ms | 1.40 |
| **full pytest** | **214 s** | **31** |

`cProfile` put **66 % of frame time in `maps.Curve.__call__`** and 75 % on the
chop -- not in the iterations. A held trim converging both pressure loops in one
pass each still cost 206 us. Every model call site asks for one abscissa and
wraps the answer in `float()`; the NumPy round-trip for that (`asarray`,
`np.all(np.isfinite(...))` at 9.4 us alone, a boolean-mask `count_nonzero`, a
`domain` property rebuilding two floats from NumPy scalars, `np.interp`'s Python
wrapper) cost 13.3 us to do about four flops.

* **`Curve._at`, a scalar path: 13.33 -> 0.93 us.** It reproduces NumPy's
  `arr_interp` rather than approximating it -- `bisect_right` for the bracket,
  `y[0]`/`y[-1]` off the ends, an exact-knot short circuit, and the slope formed
  per call because NumPy only pre-tabulates when there are more queries than
  knots. Checked over 129,694 queries across all 32 shipped curves, at every
  knot, every midpoint, 4000 random interior points each and well outside both
  ends: zero bit-mismatches. Arrays take the NumPy path unchanged.
* **`SpeedMap.__call__` never materializes the blended line: 18.81 -> 2.65 us.**
  Yesterday's `f1` beta grid had never been measured; blending two 56-point
  lines allocated two arrays a frame to serve one abscissa. The blend is now
  evaluated inside the bisection -- six knots, not 112 -- with the same
  element-wise arithmetic and so the same bits. *The 56 points were never the
  cost; materializing them was.*
* **Loop invariants out of the Gauss-Seidel sweep.** It recomputed
  `sqrt(theta41)` and `4*K_DPB*t3*wa31^2` every pass, though pdf p.37's own
  "eleven arithmetic operations for each pass" says T3 and theta41 are held.
  Hoisted with the printed expressions' association preserved. `np.sqrt` ->
  `math.sqrt` here and in `engine.frame`: same IEEE result, 1 us -> 0.05.
* **`dataclasses.replace` -> construction in `ecu`/`hmu`/`loop.step`,
  6.6 us/frame.** All three rewrite every field, so `replace` was introspecting
  the field list each frame to rediscover that it was a construction.

**Eq. 80 no longer needs a root finder.** Substituting `u = Ps9/P45` turns
`P45 = N/f9(Ps9/P45)` into `f9(u) = (N/Ps9)*u` -- a piecewise-linear function
against a line through the origin. `f9` is conditioned non-increasing and
`m = N/Ps9 > 0`, so `phi(u) = f9(u) - m*u` is strictly decreasing: the root is
unique, its segment is five comparisons away, and the root itself is one divide.
The bracketing fallback was the right answer to a repelling fixed point (#57),
but the equation never needed a *numerical* solver once the table it looks up is
the piecewise-linear thing it is.

This is the one change that is not bit-identical, and it is bounded. Over the
801 fallback solves the published transients take, the exact root satisfies
Eq. 80 to better than **1e-12** relative and the bisection's answer sits within
**3.05e-6** of it -- inside its own `TOL_P45_BISECT` of 1e-5, as it must. Only
chop runs move at all; trims, all twelve Appendix B comparisons, Figure 9, the
multirate run and the entire closed loop are bit-identical because none of them
takes this path. Within the chop, normalised by each channel's own excursion,
the worst movement in any channel is **0.00045 %** -- NG moves 0.0046 rpm in an
8188 rpm excursion. Mean `p45_iters` on the chop falls 16.90 -> 7.02.

One visible non-model change: the bisector called `f9` about eighteen times per
fallback frame and each counted a clamp. The exact solver calls it once, at the
root, so `f9`'s clamp count in the sub-idle regime falls by roughly 18x. That is
a more honest diagnostic and it is a changed number.

Verified against the pre-change tree: 71,113 values over six transients with the
heat sink both ways, a multirate run, eleven trims from 125 to 775 lbm/hr, all
5 DOF x 3 trims of `linear.extract`, and ten curves plus `f1` on dense grids
including off-domain; plus 67,500 closed-loop values across three trims with a
mid-run collective slam exercising the accel limit and all three hysteresis
blocks. Findings 1/2/3/5 together: 0 of 71,113 and 0 of 67,500 differ.

Measured as already efficient and left alone: `run()`'s trace recording (1.58
us/frame; both alternatives are slower, one was implemented and reverted),
`thermo`'s double dispatch (0.56 us/frame, and it is the C++-backend seam),
`trim.sweep` (already warm-starts), `linear.extract` FIVE (13 frame evaluations
for a 5x5 central-difference Jacobian plus the input column -- the floor), and
`_inner_pressure_loop`'s algorithm, which is the report's and stays.

+++ src/t700/engine.py
@@ -36,6 +36,7 @@ The report's, throughout: lbm, lbm/sec, psia, deg R, rpm, ft*lbf, Btu/lbm. See
 from __future__ import annotations
 
 from dataclasses import dataclass
+from math import sqrt
 
 import numpy as np
 
@@ -232,7 +233,7 @@ def frame(
     # checked claim rather than an assumption -- see
     # `tests/test_mass_conservation.py::test_the_combustor_flow_guards_never_fire`.
     dp = p3 * (p3 - p41)
-    wa31 = float(np.sqrt(max(dp, 0.0) / (c.K_DPB * t3)))  # (18)
+    wa31 = sqrt(max(dp, 0.0) / (c.K_DPB * t3))  # (18)
     substituted = dp <= 0.0 or wa31 <= 0.0
     far = wf_pps / wa31 if wa31 > 0.0 else 0.0  # (19)
     eta_b = float(maps.f6()(far))  # (20)
@@ -251,7 +252,7 @@ def frame(
     # --- gas generator turbine, Eqs. 26-28 --------------------------------------------
     dh_gt = theta41 * float(maps.f7()(p45 / p41))  # (26)
     h44 = h41 - dh_gt  # (27)
-    w41 = c.K_WGT * p41 / np.sqrt(theta41)  # (28)
+    w41 = c.K_WGT * p41 / sqrt(theta41)  # (28)
 
     # --- station 4.5 mixing, Eqs. 29-31 -----------------------------------------------
     h45 = thermo.h45_from_h44(h44)  # (29)
@@ -265,7 +266,7 @@ def frame(
     # --- power turbine, Eqs. 32-36 ----------------------------------------------------
     dh_pt = theta45 * float(maps.f8()(p49 / p45))  # (32)
     w45c = float(maps.f9()(ps9 / p45))  # (33)
-    w45 = w45c * p45 / np.sqrt(theta45)  # (34)
+    w45 = w45c * p45 / sqrt(theta45)  # (34)
     h49 = h45 - dh_pt  # (35)
     t49 = thermo.t49_from_h49(h49)  # (36)
